Divide classes into packages
Classes are currently located in the root package without clear
organization.

This unstructured organization makes the codebase harder to navigate
and maintain. It violates OOP principles of encapsulation and
modularity, making it difficult to locate related classes and
understand the overall architecture.

Divide classes into specific packages based on their functionality:
- Move command-related classes (`Command`, `AddCommand`, `ListCommand`,
  etc.) into a `command` package.
- Place task-related classes (`Task`, `ToDoTask`, `DeadlineTask`,
  `EventTask`) and exception classes (`InvalidInputException`,
  `ToDoException`, `DeadlineException`, `EventException`) into a `task`
  package.
- Organize file handling classes (`Storage`) into a `storage` package.
- Group user interaction classes (`Ui`) into a `ui` package.
- Consolidate parsing logic (`Parser`) into a `parser` package.

This organization follows OOP principles, promoting encapsulation and
modularity. It makes the codebase more readable, maintainable, and
scalable by logically grouping related functionalities. Future
development and troubleshooting will be easier, as classes will be
clearly categorized and located.

No functional changes are introduced; this refactor strictly improves
the project structure.
