+++ src/main/java/elara/command/Command.java
@@ -1,3 +1,5 @@
+package elara.command;
+
 public interface Command {
     void execute(TaskList taskList, Ui ui, Storage storage);
 }
