+++ src/main/java/elara/command/DeleteCommand.java
@@ -1,15 +1,18 @@
-public class UnmarkCommand implements Command {
+package elara.task;
+
+public class DeleteCommand implements Command {
     private final String taskDetails;
 
-    public UnmarkCommand(String taskDetails) {
+    public DeleteCommand(String taskDetails) {
         this.taskDetails = taskDetails;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         int i = Integer.parseInt(taskDetails) - 1;
-        taskList.unmarkTask(i);
-        ui.showUnmarkedTaskMessage(taskList.getTask(i));
+        ui.showRemoveTaskMessage(taskList.getTask(i));
+        taskList.deleteTask(i);
+        ui.showNumOfTasksMessage(taskList);
         storage.write(taskList);
     }
 }
