+++ src/main/java/elara/command/ExitCommand.java
@@ -1,3 +1,5 @@
+import elara.task.TaskList;
+
 public class ExitCommand implements Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
