+++ src/main/java/elara/command/UnmarkCommand.java
@@ -1,16 +1,17 @@
-public class DeleteCommand implements Command {
+import elara.task.TaskList;
+
+public class UnmarkCommand implements Command {
     private final String taskDetails;
 
-    public DeleteCommand(String taskDetails) {
+    public UnmarkCommand(String taskDetails) {
         this.taskDetails = taskDetails;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         int i = Integer.parseInt(taskDetails) - 1;
-        ui.showRemoveTaskMessage(taskList.getTask(i));
-        taskList.deleteTask(i);
-        ui.showNumOfTasksMessage(taskList);
+        taskList.unmarkTask(i);
+        ui.showUnmarkedTaskMessage(taskList.getTask(i));
         storage.write(taskList);
     }
 }
