+++ src/main/java/elara/ui/Ui.java
@@ -1,5 +1,7 @@
+import elara.task.Task;
+import elara.task.TaskList;
+
 import java.util.Scanner;
-import java.util.zip.ZipFile;
 
 public class Ui {
     private static final String LINE = "____________________________________________________________";
